refactor(tune): extract StudyConfig, TuneVerdict, create_study into study.py
Splits harness.py (317→300 lines) by moving dataclasses and study
creation into a dedicated study.py module. Updates all import sites
including CLI and test files.
